% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{AI_model_choice}
\alias{AI_model_choice}
\title{AI Model Choices}
\format{
An object of class \code{character} of length 5.
}
\usage{
AI_model_choice
}
\description{
A named character vector for AI Model keywords
}
\keyword{datasets}
