% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{custom_palette}
\alias{custom_palette}
\title{Color Choices}
\format{
An object of class \code{character} of length 6.
}
\usage{
custom_palette
}
\description{
A named character vector for color
}
\keyword{datasets}
