% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{Data_Interpretation_Prompt}
\alias{Data_Interpretation_Prompt}
\title{Data Interpretation Prompt}
\format{
An object of class \code{character} of length 1.
}
\usage{
Data_Interpretation_Prompt
}
\description{
Prompt for Data Interpretation
}
\keyword{datasets}
