% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{GT_method_choice}
\alias{GT_method_choice}
\title{Genetic Distance Choices}
\format{
An object of class \code{character} of length 8.
}
\usage{
GT_method_choice
}
\description{
A named character vector for genetic distance keywords
}
\keyword{datasets}
