% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{IBS_analysis}
\alias{IBS_analysis}
\title{IBS_analysis}
\usage{
IBS_analysis(
  data,
  Site_Info,
  REF,
  OBJ,
  Sliding.window = TRUE,
  window.size,
  step.size,
  remove_RM = TRUE
)
}
\description{
IBS_analysis
}
