% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{Idea_Expansion_Prompt}
\alias{Idea_Expansion_Prompt}
\title{Idea Expansion Prompt}
\format{
An object of class \code{character} of length 1.
}
\usage{
Idea_Expansion_Prompt
}
\description{
Prompt for Idea Expansion
}
\keyword{datasets}
