% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{Legend_choice}
\alias{Legend_choice}
\title{Tree Legend Choices}
\format{
An object of class \code{character} of length 5.
}
\usage{
Legend_choice
}
\description{
A named character vector for legend keywords
}
\keyword{datasets}
