% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{pval_adj_method_choice}
\alias{pval_adj_method_choice}
\title{P value Adjustment Method Choices}
\format{
An object of class \code{character} of length 2.
}
\usage{
pval_adj_method_choice
}
\description{
A named character vector for p value adjustment keywords
}
\keyword{datasets}
