% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{Report_Structuring_Prompt}
\alias{Report_Structuring_Prompt}
\title{Report Structuring Prompt}
\format{
An object of class \code{character} of length 1.
}
\usage{
Report_Structuring_Prompt
}
\description{
Prompt for Report Structuring
}
\keyword{datasets}
