% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{Summary_Request_Prompt}
\alias{Summary_Request_Prompt}
\title{Summary Request Prompt}
\format{
An object of class \code{character} of length 1.
}
\usage{
Summary_Request_Prompt
}
\description{
Prompt for Summary Request
}
\keyword{datasets}
