% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{Tree_layout_choice}
\alias{Tree_layout_choice}
\title{Tree Layout Choices}
\format{
An object of class \code{character} of length 8.
}
\usage{
Tree_layout_choice
}
\description{
A named character vector for layout keywords
}
\keyword{datasets}
